refactor(ops): split wide_prefill_kernel into its own module (500-LoC cap)
`prefill_attn_main_a.rs` went 481 -> 523 lines when the HDIM>256 kernel
selector landed, crossing the repo's per-file cap. CI caught it.

The helper belongs alone regardless of the line count: the kernel NAME is chosen
in `qwen3_attention::init` while the GRID is built in `prefill_attention`, two
different files, and a mismatch between them does not fail — it silently computes
the wrong q-tiles. One reader, one module.

Exact piecewise move, no behaviour change. 489 + 44 lines.
